Add a class to guard access to the LMDB database
The new class will be used to avoid passing the LMDB handle as an
argument to every function. The implementation isn't very smart so far,
but this can be improved without breaking the existing API.

The new code is located in xhalarm because it is not usable on the
client side.